% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_qc_phase1.2.R
\name{runQC_Phase1.2_report}
\alias{runQC_Phase1.2_report}
\title{Run QC}
\usage{
runQC_Phase1.2_report(dir.input, dir.output, site.nm)
}
\arguments{
\item{dir.input}{path that contains phase 1.2 data}

\item{dir.output}{path to save the QC report}

\item{site.nm}{4CE site ID (string)}
}
\value{
QC file
}
\description{
Run QC
}
